Add a file-backed bootloader interface for test purposes
When your bootloader is not supported by swupdate, it may be convenient
to have a persistent settings store put on a separate partition (not
affected by firmware updates). This is really meant to be an intermediate
solution as the hosting partition may get corrupted if the device isn't
cleanly shut down.

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>