Add SWU_OS_UPDATE and SWU_APP_UPDATE to the environment
The user-specified post-update command may need to know if the OS and/or
Applications were updated. Define SWU_OS_UPDATE and SWU_APP_UPDATE in the
environment as counters. They can be used by post-update scripts to trigger
a reboot if the OS was updated or do something else (e.g. for applications)

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>